defaulting to LW player for windows and OSx

+++ private-components/swingui/src/main/java/org/limewire/ui/swing/player/AudioPlayerMediator.java
@@ -157,7 +157,7 @@ class AudioPlayerMediator implements PlayerMediator {
      */
     @Override
     public void setVolume(double value) {
-        getPlayer().setVolume(value);
+        getPlayer().setVolume((float)value);
     }
     
     /* (non-Javadoc)
